Add BirdMember class to encapsulate bird behavior and neural network integration
